+++ src/main/java/com/example/authmodule/repository/PacienteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PacienteRepository extends JpaRepository<Paciente, Long>{
+public interface PacienteRepository extends JpaRepository<Paciente, Integer>{
 
     Paciente findPacienteByCpf(String cpf);
